feat(derive,test,spl): typed fails_with, deterministic actors, ATA and arg-seed client derivation
#[error_code] enums (QuasarError included) now convert to u32, so
fails_with takes them directly. QuasarTest fresh addresses derive from a
per-world counter instead of Pubkey::new_unique's process-global one, so
a test sees identical addresses on every run regardless of test order —
compute-unit records stay comparable without hand-pinned constants.

Client PDA derivation extends to instruction-arg seeds (Address by ref,
integer/byte-array by value) and to associated token accounts: a
behavior group named associated_token with authority + mint mapped to
plain fields derives through the behavior module's new client_address fn
(token_program joins via its canonical const when the struct pins one).
The derive stays protocol-neutral on-chain; ATA derivation is a
documented client-codegen convention.

+++ tests/programs/test-misc/src/dx_tests.rs
@@ -66,5 +66,5 @@ fn close_rejects_a_foreign_authority(q: &mut QuasarTest) {
         }
         .swap_account(intruder_pda, account),
     )
-    .fails_with(QuasarError::InvalidPda as u32);
+    .fails_with(QuasarError::InvalidPda);
 }
